+++ test/session_pool/session_pool.test.js
@@ -116,24 +116,17 @@ describe('SessionPool - testing session pool', () => {
             expect(isCalled).toBe(true); //eslint-disable-line
         });
 
-        test('should delete picked session when it is usable a create a new one', async () => {
+        test('should delete picked session when it is unusable and create a new one', async () => {
             sessionPool.maxPoolSize = 1;
-            await sessionPool.getSession();
-            const session = sessionPool.sessions[0];
+            await sessionPool.addSession();
+
+            const session = await sessionPool.getSession();
+            expect(sessionPool.sessions[0].id === session.id).toBe(true);
 
             session.errorScore += session.maxErrorScore;
-            let isCalled = false;
-            const oldRemove = sessionPool._removeSession; //eslint-disable-line
-
-            sessionPool._removeSession = (session) => { //eslint-disable-line
-                isCalled = true;
-                return oldRemove.bind(sessionPool)(session);
-            };
-
             await sessionPool.getSession();
 
-                expect(isCalled).toBe(true); //eslint-disable-line
-                expect(sessionPool.sessions[0].id === session.id).toBe(false); //eslint-disable-line
+            expect(sessionPool.sessions[0].id === session.id).toBe(false);
             expect(sessionPool.sessions).toHaveLength(1);
         });
     });
@@ -238,13 +231,17 @@ describe('SessionPool - testing session pool', () => {
         });
     });
 
-    test('should remove session', async () => {
-        for (let i = 0; i < 10; i++) {
-            await sessionPool.getSession();
-        }
-        const picked = sessionPool.getSession();
-        sessionPool._removeSession(picked); // eslint-disable-line
-        expect(sessionPool.sessions.find((s) => s.id === picked.id)).toEqual(undefined);
+    test('should remove retired sessions', async () => {
+        sessionPool.maxPoolSize = 1;
+        await sessionPool.getSession();
+
+        const session = sessionPool.sessions[0];
+        session.errorScore += session.maxErrorScore;
+        const { id: retiredSessionId } = session;
+
+        await sessionPool.getSession();
+
+        expect(sessionPool.sessions.find((s) => s.id === retiredSessionId)).toEqual(undefined);
     });
 
     test('should recreate only usable sessions', async () => {
@@ -309,5 +306,66 @@ describe('SessionPool - testing session pool', () => {
         expect(events.listenerCount(ACTOR_EVENT_NAMES_EX.PERSIST_STATE)).toEqual(1);
         await sessionPool.teardown();
         expect(events.listenerCount(ACTOR_EVENT_NAMES_EX.PERSIST_STATE)).toEqual(0);
+    });
+
+    test('should be able to create session with provided id', async () => {
+        await sessionPool.addSession({ id: 'test-session' });
+        const session = sessionPool.sessions[0];
+        expect(session.id).toBe('test-session');
+    });
+
+    test('should be able to add session instance and create new session with provided sessionOptions with addSession() ', async () => {
+        const session = new Apify.Session({ sessionPool, id: 'test-session-instance' });
+        await sessionPool.addSession(session);
+
+        await sessionPool.addSession({ id: 'test-session' });
+
+        expect(await sessionPool.getSession('test-session')).toBeDefined();
+        expect(await sessionPool.getSession('test-session-instance')).toBeDefined();
+    });
+
+    test('should not be able to add session to the pool with id already in the pool', async () => {
+        try {
+            await sessionPool.addSession({ id: 'test-session' });
+            await sessionPool.addSession({ id: 'test-session' });
+        } catch (e) {
+            expect(e.message).toBe("Cannot add session with id 'test-session' as it already exists in the pool");
+        }
+        expect.assertions(1);
+    });
+
+    test('should be able to retrieve session with provided id', async () => {
+        await sessionPool.addSession();
+        await sessionPool.addSession({ id: 'test-session' });
+        await sessionPool.addSession({ id: 'another-test-session' });
+
+        const session = await sessionPool.getSession('test-session');
+        expect(session.id).toBe('test-session');
+    });
+
+    test('should correctly populate session array and session map', async () => {
+        sessionPool.maxPoolSize = 10;
+
+        for (let i = 0; i < 20; i++) await sessionPool.getSession();
+
+        expect(sessionPool.sessions.length).toEqual(10);
+        expect(sessionPool.sessionMap.size).toEqual(10);
+        expect(sessionPool.sessions.length).toEqual(sessionPool.sessionMap.size);
+    });
+
+    test('should correctly remove retired sessions both from array and session map', async () => {
+        sessionPool.maxPoolSize = 10;
+
+        for (let i = 0; i < 10; i++) {
+            await sessionPool.addSession({ id: `session_${i}` });
+            const session = await sessionPool.getSession(`session_${i}`);
+            session.errorScore += session.maxErrorScore;
+        }
+
+        await sessionPool.getSession();
+
+        expect(sessionPool.sessions.length).toEqual(1);
+        expect(sessionPool.sessionMap.size).toEqual(1);
+        expect(sessionPool.sessions.length).toEqual(sessionPool.sessionMap.size);
     });
 });